+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## CHANGELOG
 
-### 3.0.0 (19-Feb-2026)
+### 3.1.0 (03-Mar-2026)
+- Refactored code to more reliably display `$humanReadableScriptName` in the Dock
+- Added Volume Owners to `$helpmessage`
+
+### 3.0.0 (23-Feb-2026)
 **First (attempt at a) MDM-agnostic release**
 - Added a new `Development` Operation Mode to aid in developing / testing individual Health Checks. (See: [README.md](README.md) for details.)
 - Minor update to host check curl logic (Pull Request #60; thanks, @ecubrooks!)